[test] auth 모듈 e2e 테스트 추가

+++ src/config/database/database.module.ts
@@ -8,18 +8,35 @@ import { ConfigService } from '../config.service';
     TypeOrmModule.forRootAsync({
       imports: [ConfigModule],
       inject: [ConfigService],
-      useFactory: (configService: ConfigService) => ({
-        type: 'mysql',
-        host: configService.get('DATABASE_HOST'),
-        port: Number(configService.get('DATABASE_PORT')),
-        username: configService.get('DATABASE_USER'),
-        password: configService.get('DATABASE_PASSWORD'),
-        database: configService.get('DATABASE_NAME'),
-        entities: [__dirname + './../../**/**.model{.ts,.js}'],
-        synchronize: false,
-        dropSchema: false,
-        logging: false,
-      }),
+      useFactory: (configService: ConfigService) => {
+        if (process.env.NODE_ENV === 'test') {
+          return {
+            type: 'mysql',
+            host: '127.0.0.1',
+            port: 3306,
+            username: configService.get('DATABASE_USER'),
+            password: configService.get('DATABASE_PASSWORD'),
+            database: configService.get('DATABASE_NAME'),
+            entities: [__dirname + './../../**/**.model{.ts,.js}'],
+            synchronize: false,
+            dropSchema: false,
+            logging: false,
+          };
+        }
+
+        return {
+          type: 'mysql',
+          host: configService.get('DATABASE_HOST'),
+          port: Number(configService.get('DATABASE_PORT')),
+          username: configService.get('DATABASE_USER'),
+          password: configService.get('DATABASE_PASSWORD'),
+          database: configService.get('DATABASE_NAME'),
+          entities: [__dirname + './../../**/**.model{.ts,.js}'],
+          synchronize: false,
+          dropSchema: false,
+          logging: false,
+        };
+      },
     }),
   ],
 })
